# docs/swagger.yaml
basePath: /
definitions:
...
  handler.PaymentRequest:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
...
host: localhost:8080
info:
  contact:
    email: Obie.kal22@gmail.com
    url: www.linkedin.com/in/obie-ananda-a87a64212
  description: API documentation for the FTGO PlashCash project.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: FTGO PlasCash Project
  version: "1.0"
paths:
  /customers/purchase-status/{order_id}:
    get:
      consumes:
      - application/json
      description: Checks the status of a customer's purchase transaction and updates
        inventory if successful.
      parameters:
      - description: Order ID of the transaction
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Purchase status and inventory updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid transaction or insufficient stock
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch transaction status or update inventory
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check purchase status
      tags:
      - Customers
  /customers/tokens:
    get:
      consumes:
      - application/json
      description: Retrieves all tokens for the logged-in customer.
      produces:
      - application/json
      responses:
        "200":
          description: Customer tokens fetched successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Missing or invalid customer ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch or parse tokens
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Retrieve customer tokens
      tags:
      - Customers
  /customers/wallet-balance:
    get:
      consumes:
      - application/json
      description: Retrieves the wallet balance for the logged-in customer.
      produces:
      - application/json
      responses:
        "200":
          description: Wallet balance retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve wallet balance
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Retrieve customer wallet balance
      tags:
      - Customers
  /customers/withdraw:
    post:
      consumes:
      - application/json
      description: Processes a withdrawal request for the customer and creates a virtual
        account number for the transaction.
      parameters:
      - description: Withdrawal request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Withdrawal initiated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or withdrawal amount
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to process withdrawal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Initiate a withdrawal for the customer
      tags:
      - Customers
  /customers/withdrawal-status/{order_id}:
    get:
      consumes:
      - application/json
      description: Checks the status of a customer's withdrawal transaction.
      parameters:
      - description: Order ID of the transaction
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction status retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid transaction request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch transaction status
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check withdrawal status
      tags:
      - Customers
...
  /stores/coordinates:
    get:
      consumes:
      - application/json
      description: Retrieves the coordinates of all stores.
      produces:
      - application/json
      responses:
        "200":
          description: Store coordinates retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch store coordinates
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve all store coordinates
      tags:
      - Stores
...
swagger: "2.0"